requirements.txt
azure-storage-blob
flask
flask-restful